test(activerecord): use decimal for composite-key float-scope test column
MariaDB FLOAT loses precision for 9.99, causing the composite-key
float-scope uniqueness lookup to miss the existing row.

+++ packages/activerecord/src/validations/uniqueness-validation.test.ts
@@ -887,7 +887,7 @@ async function freshAdapterCompositeOrders(): Promise<DatabaseAdapter> {
       c: "integer",
       val: "string",
       active: "boolean",
-      price: "float",
+      price: "decimal",
     },
     order_as: { shop_id: "integer", order_num: "integer" },
     order_bs: { shop_id: "integer", order_num: "integer" },
